Added recognition of SSH authentication errors for interactive nodes
If new nodes are added to the cluster, or configuration is messed up, SSH authentication errors may occur when connecting to the interactive node while starting an interactive session. With these changes, Jupyter-O2 should at least alert the user and close if such errors are found, rather than hanging silently.

Added reactions to FilteredOut to allow closing Jupyter-O2 on encountering the authentication error.