Implement Duo two-factor authentication handler
Adds two command-line arguments, --2fa and --2fa-code, which enable 2FA
handling and provide code(s) for 2FA, respectively. The 2FA code(s) can
be the number 1 to automatically trigger a Duo Push request, or pre-
generated 2FA codes can be entered. If no code is provided, an
interactive session is started to enable user interaction with the Duo
prompt.

Both command-line arguments have corresponding config options
if they are to be enabled constantly. However, note that there is not
yet any argument that overrides the config and disables 2FA, so logins
may hang if 2FA is enabled in the config and the server never requests
2FA.